refactor: move email confirmation in helpers
